requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
dotenv-python==0.0.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.38
langchain-openai==0.1.23
langsmith==0.1.116
numpy==1.24.4
openai==1.44.0
orjson==3.10.7
packaging==24.1
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
sniffio==1.3.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
